new paradigm and reorganization
